+++ WorkoutManager.BusinessLogic/DTOs/WorkoutPlanDtos.cs
@@ -15,46 +15,46 @@ public record CreateWorkoutPlanDto
 }
 
 public record WorkoutPlanDto(
-    int Id,
+    long Id,
     string Name,
     DateTime CreatedAt
 );
 
 public record CreatedTrainingDayDto(
-    int Id,
+    long Id,
     string Name,
     int Order
 );
 
 public record CreatedWorkoutPlanDto(
-    int Id,
+    long Id,
     string Name,
     IEnumerable<CreatedTrainingDayDto> TrainingDays
 );
 
 public record CreatedPlanDayExerciseDto(
-    int Id,
-    int TrainingDayId,
-    int ExerciseId,
+    long Id,
+    long TrainingDayId,
+    long ExerciseId,
     int Order
 );
 
 public record PlanDayExerciseDto(
-    int PlanDayExerciseId,
-    int ExerciseId,
+    long PlanDayExerciseId,
+    long ExerciseId,
     string Name,
     int Order
 );
 
 public record TrainingDayDto(
-    int Id,
+    long Id,
     string Name,
     int Order,
     IEnumerable<PlanDayExerciseDto> Exercises
 );
 
 public record WorkoutPlanDetailDto(
-    int Id,
+    long Id,
     string Name,
     bool IsLocked,
     IEnumerable<TrainingDayDto> TrainingDays
